Fix single session

+++ HemaDungeon/Controllers/FightController.cs
@@ -1,6 +1,7 @@
 ﻿using HemaDungeon.Entities;
 using HemaDungeon.Models;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -12,17 +13,20 @@ public sealed class FightController : ControllerBase
 {
     [HttpGet("users")]
     [Authorize]
-    public async Task<IActionResult> GetUser([FromServices] Context context)
+    public async Task<IActionResult> GetUser([FromServices] Context context, [FromServices] UserManager<Character> manager)
     {
-        var users = await context.FightCharacters.Include(x => x.Character).ToListAsync();
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+        var users = await context.FightCharacters.Where(x => x.AuthorId == userId).Include(x => x.Character).ToListAsync();
         return new JsonResult(users);
     }
 
     [HttpPost("users")]
     [Authorize]
-    public async Task<IActionResult> PostUser([FromForm] FightUsersModel model, [FromServices] Context context)
+    public async Task<IActionResult> PostUser([FromForm] FightUsersModel model, [FromServices] UserManager<Character> manager, [FromServices] Context context)
     {
-        context.FightCharacters.RemoveRange(context.FightCharacters.ToList());
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+
+        context.FightCharacters.RemoveRange(context.FightCharacters.Where(x => x.AuthorId == userId).ToList());
         await context.SaveChangesAsync();
 
         foreach (var user in context.Users.Where(x => model.Names.Contains(x.Name)))
@@ -31,6 +35,7 @@ public sealed class FightController : ControllerBase
             {
                 Id = Guid.NewGuid().ToString(),
                 Character = user,
+                AuthorId = userId,
                 Health = user.Vitality
             });
         }
@@ -41,19 +46,23 @@ public sealed class FightController : ControllerBase
 
     [HttpPost("complete")]
     [Authorize]
-    public async Task<IActionResult> Complete([FromServices] Context context)
+    public async Task<IActionResult> Complete([FromServices] Context context, [FromServices] UserManager<Character> manager)
     {
-        context.FightStates.RemoveRange(context.FightStates.ToList());
-        context.FightCharacters.RemoveRange(context.FightCharacters.ToList());
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+
+        context.FightStates.RemoveRange(context.FightStates.Where(x => x.AuthorId == userId).ToList());
+        context.FightCharacters.RemoveRange(context.FightCharacters.Where(x => x.AuthorId == userId).ToList());
         await context.SaveChangesAsync();
         return Redirect("/");
     }
 
     [HttpPost("state/complete")]
     [Authorize]
-    public async Task<IActionResult> CompleteState([FromForm] FightStateModel model, [FromServices] Context context)
+    public async Task<IActionResult> CompleteState([FromForm] FightStateModel model, [FromServices] Context context, [FromServices] UserManager<Character> manager)
     {
-        var states = await context.FightStates.Include(x => x.Character).ThenInclude(x => x.Character).ToListAsync();
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+
+        var states = await context.FightStates.Where(x => x.AuthorId == userId).Include(x => x.Character).ThenInclude(x => x.Character).ToListAsync();
         states[0].Character.Health -= states[1].Damage * (model.Score[1] ?? 0);
         if (states[0].Character.Health < 0) states[0].Character.Health = 0;
 
@@ -69,18 +78,21 @@ public sealed class FightController : ControllerBase
 
     [HttpPost("state")]
     [Authorize]
-    public async Task<IActionResult> PostState([FromForm] FightUsersModel model, [FromServices] Context context)
+    public async Task<IActionResult> PostState([FromForm] FightUsersModel model, [FromServices] UserManager<Character> manager, [FromServices] Context context)
     {
-        context.FightStates.RemoveRange(context.FightStates.ToList());
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+
+        context.FightStates.RemoveRange(context.FightStates.Where(x => x.AuthorId == userId).ToList());
         await context.SaveChangesAsync();
 
         var states = new List<FightState>();
-        foreach (var user in context.FightCharacters.Include(x => x.Character).Include(x => x.Character).Where(x => model.Names.Contains(x.Character.Name)))
+        foreach (var user in context.FightCharacters.Where(x => x.AuthorId == userId).Include(x => x.Character).Include(x => x.Character).Where(x => model.Names.Contains(x.Character.Name)))
         {
             var state = new FightState
             {
                 Id = Guid.NewGuid().ToString(),
                 Character = user,
+                AuthorId = userId,
             };
             states.Add(state);
 
@@ -132,9 +144,11 @@ public sealed class FightController : ControllerBase
 
     [HttpGet("state")]
     [Authorize]
-    public async Task<IActionResult> GetState([FromServices] Context context)
+    public async Task<IActionResult> GetState([FromServices] Context context, [FromServices] UserManager<Character> manager)
     {
-        var states = await context.FightStates.Include(x => x.Character).ThenInclude(x => x.Character).ToListAsync();
+        var userId = (await manager.GetUserAsync(HttpContext.User))?.Id;
+
+        var states = await context.FightStates.Where(x => x.AuthorId == userId).Include(x => x.Character).ThenInclude(x => x.Character).ToListAsync();
         return new JsonResult(states);
     }
 }
